resume_sorting_tool/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
Django==5.1.4
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
numpy==2.0.2
packaging==24.2
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.4.8
thinc==8.3.2
tika==2.6.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
